Exercise 4.21* done
Blogilistan laajennus, step 9

+++ osa4/blogilista/controllers/blogs.js
@@ -49,6 +49,12 @@ blogsRouter.post('/', async (request, response) => {
 })
   
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(400).json({ error: 'Cannot delete a blog added by another user' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
